+++ R/convert_to_sparse.R
@@ -1,0 +1,8 @@
+convert_to_sparse <- function(M) {
+
+  methods::as(methods::as(methods::as(M,
+           "dMatrix"),
+        "generalMatrix"),
+     "CsparseMatrix")
+
+}